myapp/requirements.txt
Django==2.2
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
